+++ backend/book-service/src/main/java/kr/mybrary/bookservice/booksearch/domain/dto/kakaoapi/Meta.java
@@ -1,4 +1,4 @@
-package kr.mybrary.bookservice.book.domain.dto.kakaoapi;
+package kr.mybrary.bookservice.booksearch.domain.dto.kakaoapi;
 
 import lombok.Data;
